feat/refactor_animation(ref): change duration from float to unit32_t the unit is now a game cycle

+++ src/Animations/Explosion.hh
@@ -14,7 +14,7 @@
 
 class Explosion : public IEffect {
 public:
-    explicit Explosion(irr::core::vector3df const & pos, float duration = 1.0f);
+    explicit Explosion(irr::core::vector3df const & pos, uint32_t duration = 1);
 
     Explosion(Explosion const & other);
     Explosion(Explosion && other);
